requirements.txt
tavily-python
requests
google-generativeai